requirements.txt
PyQt5>=5.15
pyqtgraph>=0.13
PyOpenGL>=3.1
numpy>=1.24
pandas>=2.0
scipy>=1.10
pyarrow>=14.0
alpaca-py>=0.30
requests>=2.31
python-dateutil>=2.8
